+++ src/app/settings/settings.module.ts
@@ -3,13 +3,14 @@ import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatIconModule } from '@angular/material/icon';
-import { MatToolbarModule } from '@angular/material/toolbar';
 
 import { SettingsRoutingModule } from './settings-routing.module';
 import { SettingsComponent } from './settings.component';
 
+import { ToolbarModule } from '~components/toolbar/toolbar.module';
+
 @NgModule({
   declarations: [SettingsComponent],
-  imports: [CommonModule, MatButtonModule, MatDividerModule, MatIconModule, MatToolbarModule, SettingsRoutingModule]
+  imports: [CommonModule, MatButtonModule, MatDividerModule, MatIconModule, SettingsRoutingModule, ToolbarModule]
 })
 export class SettingsModule {}